% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.sim.R
\name{run.sim}
\alias{run.sim}
\title{A Unified Platform for Comprehensive Evaluation of Phase I Clinical Trial Designs}
\usage{
run.sim(
  prob,
  nSim = 10,
  dose = 3^{     0:(length(prob) - 1) },
  ttl = 0.3,
  n = 60,
  m = 3,
  Max_FoldChange = 3,
  epsilon1 = 0.05,
  epsilon2 = 0.05,
  pod1 = 0.95,
  alpha = 1,
  beta = 1,
  cohort.size = c(3, 5, 8, 10),
  theta = 0.25,
  alpha1 = 1,
  beta1 = 4
)
}
\arguments{
\item{prob}{Vector of true DLT rates}

\item{nSim}{Number of studies to simulate}

\item{dose}{Dose levels of recruited patients}

\item{ttl}{Target toxicity level}

\item{n}{The maximum sample size of the trial}

\item{m}{The cohort size}

\item{Max_FoldChange}{Maximum fold change}

\item{pod1}{Eliminate the current dose and higher if the posterior probability of exceeding the target toxicity level exceeds this value}

\item{alpha}{Prior shape1 parameter}

\item{beta}{Prior shape2 parameter}

\item{cohort.size}{Vector of cohort sizes}

\item{theta}{Safety threshold for cohort-sequence design}

\item{alpha1}{Prior shape1 parameter for cohort-sequence design}

\item{beta1}{Prior shape2 parameter for cohort-sequence design}
}
\value{
\code{run.sim( )} returns a list of simulation results
}
\description{
This is a description
}
\examples{
run.sim(prob=c(0.21, 0.30, 0.37, 0.43, 0.48, 0.52, 0.56, 0.61, 0.64, 0.66))

}
